+++ examples/character.eg.ts
@@ -3,7 +3,6 @@ import "@std/dotenv/load"
 import { T } from "structured-outputs"
 import { ResponseFormat } from "structured-outputs/client"
 import { dbg } from "testing"
-import * as std from "../json/combinators/derived/mod.ts"
 import { asserts } from "../util/mod.ts"
 
 const greeting = T.taggedUnion("greeting", {
@@ -16,16 +15,16 @@ const Character = T.object({
   name: T.string.assert(asserts.string.minLength, 4).assert(asserts.string.maxLength, 30),
   home: T.string`The name of a fictional realm of magic and wonder.`,
   disposition: T.enum("Optimistic", "Reserved", "Inquisitive"),
-  born: std.Date`Date the character was born. Make sure it aligns with the age.`,
-  stateOfAffairs: std.Tuple(
+  born: T.Date`Date the character was born. Make sure it aligns with the age.`,
+  stateOfAffairs: T.Tuple(
     T.string`Home life.`,
     T.string`Professional life.`,
     T.string`Health.`,
   )`How are things going for the character in these various domains?`,
-  randomValue: std.Union(T.string, T.number),
+  randomValue: T.Union(T.string, T.number),
   friends: T.array(T.string)`Names of the character's friends.`,
   greeting,
-  favoriteColor: std.colors.Hex,
+  favoriteColor: T.colors.Hex,
 })
 
 const response_format = ResponseFormat("create_character", Character)`
