--- json/combinators/boolean.ts
@@ -1,5 +1,0 @@
-import { Type } from "../../core/mod.ts"
-
-export const boolean: Type<boolean> = Type({
-  getType: () => boolean,
-})
