--- json/combinators/mod.ts
@@ -1,12 +1,0 @@
-// moderate
-
-export * from "./array.ts"
-export * from "./boolean.ts"
-export * from "./derived/mod.ts"
-export * from "./enum.ts"
-export * from "./number.ts"
-export * from "./object.ts"
-export * from "./option.ts"
-export * from "./string.ts"
-export * from "./taggedUnion.ts"
-export * from "./transform.ts"
