--- json/combinators/number.ts
@@ -1,5 +1,0 @@
-import { Type } from "../../core/mod.ts"
-
-export const number: Type<number> = Type({
-  getType: () => number,
-})
