--- json/combinators/string.ts
@@ -1,5 +1,0 @@
-import { Type } from "../../core/mod.ts"
-
-export const string: Type<string> = Type<string>({
-  getType: () => string,
-})
