+++ json/mod.ts
@@ -1,6 +1,6 @@
-export * as T from "./combinators/mod.ts"
+export * as T from "./T.ts"
 
-// moderate --exclude combinators
+// moderate --exclude combinators.ts derived T.ts
 
 export * from "./deserialize.ts"
 export * from "./Diagnostic.ts"
